fix: change search query from updated:> to created:> to find newly created items
The search was using 'updated:>' which only finds items updated in the last 24 hours.
This missed newly created items like Issue #132 that haven't been updated yet.

Changed to 'created:>' to find items created in the last 24 hours, which will
include newly created issues and pull requests that should be added to the project board.